src/agent2/requirements.txt
streamlit
dask
pandas
numpy
scikit-learn
plotly
altair
statsmodels
pyarrow
tensorflow
umap-learn
scipy
matplotlib
seaborn
watchdog
